backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.23.2
torch==2.6.1
torchvision==0.15.2
timm==0.9.1
huggingface-hub==0.17.3
pillow==10.0.1
pytest==8.2.3
httpx==0.26.2
pytest-asyncio==0.21.0
python-multipart==0.0.6
aiofiles==23.1.0

